+++ weindependent-app/src/main/java/com/weindependent/app/database/mapper/weindependent/EventMapper.java
@@ -1,7 +1,7 @@
 package com.weindependent.app.database.mapper.weindependent;
 
-import com.weindependent.app.vo.EventVO;
-import com.weindependent.app.vo.RecentEventVO;
+import com.weindependent.app.vo.event.EventVO;
+import com.weindependent.app.vo.event.RecentEventVO;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -10,5 +10,5 @@ import java.util.List;
 @Mapper
 public interface EventMapper {
     List<RecentEventVO> getRecent(@Param("size") int size, @Param("offset") int offset);
-    EventVO getById(Integer id);
+    EventVO getById(Long id);
 }
